+++ documentation/source/CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
 - ability to override the group based on dirname, e.g. `run_suite($component_id, 'group_alias');`
 - the log file should not be deleted, only truncated between runners.
 - "why" does not work with "import", it needs to be able to be there and override as well.
@@ -32,6 +33,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### Changed
 
+- BREAKING CHANGE! The _drupal_ mixin will now remove all suite variables it added at the end of the test. If you want to use them across tests you need to use `set` on the test.
 - _phpstorm.http_ mixin name changed to _http_request_files_; update any `add_mixin()` calls that reference this.
 - Breakpoints will run only when passing `--break`; previously they ran in verbose mode.
 - `expect` changed to `status`; update all tests.